+++ src/apps/weblib/js-api/vizzu.js
@@ -396,6 +396,10 @@ export default class Vizzu {
       this.call(this.module._vizzu_mouseLeave)();
     });
 
+    canvas.addEventListener("wheel", (evt) => {
+      this.call(this.module._vizzu_mousewheel)(evt.deltaY);
+    });
+
     document.addEventListener("keydown", (evt) => {
       let key = evt.keyCode <= 255 ? evt.keyCode : 0;
       const keys = [33, 34, 36, 35, 37, 39, 38, 40, 27, 9, 13, 46];